pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mma-pipeline"
version = "0.1.0"
description = "Modular Multi-task Architecture for MMA Athlete Tracking and Pose Estimation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "MMA Project Team", email = ""},
]
keywords = ["tracking", "pose-estimation", "YOLO", "SMPL", "athlete-tracking"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "opencv-python>=4.5.0",
    "numpy>=1.21.0",
    "pillow>=8.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ultralytics>=8.0.0",
    "torchreid>=1.3.0",
    "scipy>=1.7.0",
    "filterpy>=1.4.5",
    "pandas>=1.3.0",
    "pyyaml>=5.4.0",
    "tqdm>=4.60.0",
    "matplotlib>=3.4.0",
    "trimesh>=3.9.0",
    "pyrender>=0.1.45",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.0",
    "pytest-cov>=2.12.0",
    "black>=21.0",
    "isort>=5.9.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
]
mmpose = [
    "mmcv>=1.4.0",
    "mmpose>=0.25.0",
    "mmdet>=2.20.0",
    "mmengine>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/your-org/mma-pipeline"
Documentation = "https://github.com/your-org/mma-pipeline/wiki"
Repository = "https://github.com/your-org/mma-pipeline"
"Bug Tracker" = "https://github.com/your-org/mma-pipeline/issues"

[tool.setuptools]
packages = ["mma"]

[tool.setuptools.package-data]
mma = ["config/*.yaml"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers -v"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["mma"]
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*Protocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
